@@ -4,8 +4,8 @@ apply plugin: 'com.github.triplet.play'
 
 dependencies {
     implementation 'eu.chainfire:libsuperuser:1.1.0.201903290044'
-    implementation 'com.android.support:design:28.0.0'
-    implementation 'com.android.support:recyclerview-v7:28.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
     implementation 'com.google.zxing:android-integration:3.3.0'
     implementation 'com.google.code.gson:gson:2.8.5'
     implementation 'org.mindrot:jbcrypt:0.4'
@@ -13,11 +13,11 @@ dependencies {
     implementation 'com.google.guava:guava:27.1-android'
     implementation 'com.annimon:stream:1.2.1'
     implementation 'com.android.volley:volley:1.1.1'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     implementation 'com.google.dagger:dagger:2.23.1'
     annotationProcessor "com.google.dagger:dagger-compiler:2.23.1"
-    androidTestImplementation 'com.android.support.test:rules:1.0.2'
-    androidTestImplementation 'com.android.support:support-annotations:28.0.0'
+    androidTestImplementation 'androidx.test:rules:1.1.0'
+    androidTestImplementation 'androidx.annotation:annotation:1.0.0'
 }
 
 apply from: 'versions.gradle'
@@ -26,8 +26,8 @@ def ourVersionName = "${versionMajor}.${versionMinor}.${versionPatch}.${versionW
 
 android {
     // Changes to these values need to be reflected in `.travis.yml`
-    compileSdkVersion 28
-    buildToolsVersion '28.0.3'
+    compileSdkVersion 29
+    buildToolsVersion '29.0.0'
 
     buildTypes.debug.applicationIdSuffix ".debug"
     dataBinding.enabled = true
@@ -40,11 +40,11 @@ android {
     defaultConfig {
         applicationId "com.github.catfriend1.syncthingandroid"
         minSdkVersion 16
-        targetSdkVersion 28
+        targetSdkVersion 29
         versionCode ourVersionCode
         versionName ourVersionName
         testApplicationId 'com.github.catfriend1.syncthingandroid.test'
-        testInstrumentationRunner 'android.support.test.runner.AndroidJUnitRunner'
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
 
     signingConfigs {
