@@ -3,8 +3,8 @@ package com.nutomic.syncthingandroid.activities;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.app.AppCompatDelegate;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatDelegate;
 // import android.util.Log;
 
 import com.nutomic.syncthingandroid.service.Constants;
