@@ -4,7 +4,7 @@ import android.annotation.TargetApi;
 import android.annotation.SuppressLint;
 import android.content.Context;
 import android.content.res.Resources;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.net.Uri;
 import android.os.Build;
 // import android.util.Log;
