@@ -1,7 +1,7 @@
 package com.nutomic.syncthingandroid.views;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 // import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
